Do not import from itemadapter outside a test method
