+++ tests/test_adapter.py
@@ -1,14 +1,11 @@
 import unittest
+from importlib import import_module
 from types import MappingProxyType
 from typing import KeysView
 
-from itemadapter.adapter import ItemAdapter
-
 from tests import (
     AttrsItem,
-    AttrsItemNested,
     DataClassItem,
-    DataClassItemNested,
     requires_attr,
     requires_dataclasses,
     requires_scrapy,
@@ -20,6 +17,8 @@ from tests import (
 
 class ItemAdapterReprTestCase(TestCase):
     def test_repr_dict(self):
+        from itemadapter.adapter import ItemAdapter
+
         item = dict(name="asdf", value=1234)
         adapter = ItemAdapter(item)
         # dicts are not guarantied to be sorted in py35
@@ -30,6 +29,8 @@ class ItemAdapterReprTestCase(TestCase):
 
     @requires_scrapy
     def test_repr_scrapy_item(self):
+        from itemadapter.adapter import ItemAdapter
+
         item = ScrapySubclassedItem(name="asdf", value=1234)
         adapter = ItemAdapter(item)
         # Scrapy fields are stored in a dict, which is not guarantied to be sorted in py35
@@ -40,6 +41,8 @@ class ItemAdapterReprTestCase(TestCase):
 
     @requires_dataclasses
     def test_repr_dataclass(self):
+        from itemadapter.adapter import ItemAdapter
+
         item = DataClassItem(name="asdf", value=1234)
         adapter = ItemAdapter(item)
         self.assertEqual(
@@ -48,6 +51,8 @@ class ItemAdapterReprTestCase(TestCase):
 
     @requires_attr
     def test_repr_attrs(self):
+        from itemadapter.adapter import ItemAdapter
+
         item = AttrsItem(name="asdf", value=1234)
         adapter = ItemAdapter(item)
         self.assertEqual(
@@ -57,6 +62,8 @@ class ItemAdapterReprTestCase(TestCase):
 
 class ItemAdapterInitError(TestCase):
     def test_non_item(self):
+        from itemadapter.adapter import ItemAdapter
+
         with self.assertRaises(TypeError):
             ItemAdapter(ScrapySubclassedItem)
         with self.assertRaises(TypeError):
@@ -68,14 +75,22 @@ class ItemAdapterInitError(TestCase):
 class BaseTestMixin:
 
     item_class = None
-    item_class_nested = None
+    item_class_nested_path = None
 
     def setUp(self):
         super().setUp()
         if self.item_class is None:
             raise unittest.SkipTest()
 
+    @property
+    def item_class_nested(self):
+        module_path, class_name = self.item_class_nested_path.rsplit(".", maxsplit=1)
+        module = import_module(module_path)
+        return getattr(module, class_name)
+
     def test_get_set_value(self):
+        from itemadapter.adapter import ItemAdapter
+
         item = self.item_class()
         adapter = ItemAdapter(item)
         self.assertEqual(adapter.get("name"), None)
@@ -95,17 +110,23 @@ class BaseTestMixin:
         self.assertEqual(adapter["value"], 1234)
 
     def test_get_value_keyerror(self):
+        from itemadapter.adapter import ItemAdapter
+
         item = self.item_class()
         adapter = ItemAdapter(item)
         with self.assertRaises(KeyError):
             adapter["undefined_field"]
 
     def test_as_dict(self):
+        from itemadapter.adapter import ItemAdapter
+
         item = self.item_class(name="asdf", value=1234)
         adapter = ItemAdapter(item)
         self.assertEqual(dict(name="asdf", value=1234), dict(adapter))
 
     def test_as_dict_nested(self):
+        from itemadapter.adapter import ItemAdapter
+
         item = self.item_class_nested(
             nested=self.item_class(name="asdf", value=1234),
             adapter=ItemAdapter(dict(foo="bar", nested_list=[1, 2, 3, 4, 5])),
@@ -130,6 +151,8 @@ class BaseTestMixin:
         )
 
     def test_field_names(self):
+        from itemadapter.adapter import ItemAdapter
+
         item = self.item_class(name="asdf", value=1234)
         adapter = ItemAdapter(item)
         self.assertIsInstance(adapter.field_names(), KeysView)
@@ -138,12 +161,16 @@ class BaseTestMixin:
 
 class NonDictTestMixin(BaseTestMixin):
     def test_set_value_keyerror(self):
+        from itemadapter.adapter import ItemAdapter
+
         item = self.item_class()
         adapter = ItemAdapter(item)
         with self.assertRaises(KeyError):
             adapter["undefined_field"] = "some value"
 
     def test_metadata_common(self):
+        from itemadapter.adapter import ItemAdapter
+
         adapter = ItemAdapter(self.item_class())
         self.assertIsInstance(adapter.get_field_meta("name"), MappingProxyType)
         self.assertIsInstance(adapter.get_field_meta("value"), MappingProxyType)
@@ -151,11 +178,15 @@ class NonDictTestMixin(BaseTestMixin):
             adapter.get_field_meta("undefined_field")
 
     def test_get_field_meta_defined_fields(self):
+        from itemadapter.adapter import ItemAdapter
+
         adapter = ItemAdapter(self.item_class())
         self.assertEqual(adapter.get_field_meta("name"), MappingProxyType({"serializer": str}))
         self.assertEqual(adapter.get_field_meta("value"), MappingProxyType({"serializer": int}))
 
     def test_delitem_len_iter(self):
+        from itemadapter.adapter import ItemAdapter
+
         item = self.item_class(name="asdf", value=1234)
         adapter = ItemAdapter(item)
         self.assertEqual(len(adapter), 2)
@@ -184,16 +215,22 @@ class DictTestCase(TestCase, BaseTestMixin):
 
     def test_get_value_keyerror_item_dict(self):
         """Instantiate without default values"""
+        from itemadapter.adapter import ItemAdapter
+
         adapter = ItemAdapter(self.item_class())
         with self.assertRaises(KeyError):
             adapter["name"]
 
     def test_empty_metadata(self):
+        from itemadapter.adapter import ItemAdapter
+
         adapter = ItemAdapter(self.item_class(name="foo", value=5))
         for field_name in ("name", "value", "undefined_field"):
             self.assertEqual(adapter.get_field_meta(field_name), MappingProxyType({}))
 
     def test_field_names_updated(self):
+        from itemadapter.adapter import ItemAdapter
+
         item = self.item_class(name="asdf")
         field_names = ItemAdapter(item).field_names()
         self.assertEqual(sorted(field_names), ["name"])
@@ -208,6 +245,8 @@ class ScrapySubclassedItemTestCase(NonDictTestMixin, TestCase):
 
     def test_get_value_keyerror_item_dict(self):
         """Instantiate without default values"""
+        from itemadapter.adapter import ItemAdapter
+
         adapter = ItemAdapter(self.item_class())
         with self.assertRaises(KeyError):
             adapter["name"]
@@ -216,10 +255,10 @@ class ScrapySubclassedItemTestCase(NonDictTestMixin, TestCase):
 class DataClassItemTestCase(NonDictTestMixin, TestCase):
 
     item_class = DataClassItem
-    item_class_nested = DataClassItemNested
+    item_class_nested_path = "tests.dataclasses_utils.DataClassItemNested"
 
 
 class AttrsItemTestCase(NonDictTestMixin, TestCase):
 
     item_class = AttrsItem
-    item_class_nested = AttrsItemNested
+    item_class_nested_path = "tests.attr_utils.AttrsItemNested"
